remove "branches:" from on.pull_requests
I found having `on.pull_requests: branches [ $default-branch ]` causes PR's against the default branch NOT to trigger a build.
Removing `branches: [ $default-branch ]` from below `pull_request:` allows a build to proceed when a PR against the default branch is created.
Apologies if I misinterpret the intent of the original config, but it seemed like a copy/pasta error.
Or perhaps the intent is only to trigger builds in Forks? If so, it might be work clarifying that in the docs.